refactor: 관심사에 따른 타입 선언 위치 분리
