+++ frontend/src/api/get.ts
@@ -1,28 +1,30 @@
 import { api, customerHeader, ownerHeader } from '.';
 import { PARAMS_ERROR_MESSAGE } from '../constants';
-import { CouponDesign } from '../types';
 import {
+  QueryReq,
+  PhoneNumberParams,
   CafeIdParams,
-  CafeInfoRes,
-  CafeRes,
-  CouponRes,
   CustomerIdParams,
+  MaxStampCountParams,
+  UsedParams,
+  OAuthTokenParams,
+  CouponIdParams,
+} from '../types/api/request';
+import {
+  CafeRes,
   CustomerPhoneNumberRes,
   CustomersRes,
   IssuedCouponsRes,
-  MaxStampCountParams,
-  OAuthJWTRes,
-  OAuthTokenParams,
-  MyRewardRes,
-  PhoneNumberParams,
-  QueryReq,
-  RewardRes,
   SampleCouponRes,
+  CouponRes,
+  CafeInfoRes,
+  MyRewardRes,
   StampHistoryRes,
-  UsedParams,
+  OAuthJWTRes,
   CustomerProfileRes,
-  CouponIdParams,
-} from '../types/api';
+  RewardRes,
+} from '../types/api/response';
+import { CouponDesign } from '../types/domain/coupon';
 
 export const getCafe = async () => {
   return await api.get<CafeRes>('/admin/cafes', ownerHeader());
