+++ frontend/src/components/Template/CustomerTemplate/BottomTabBar/index.tsx
@@ -2,7 +2,7 @@ import { Link, useLocation } from 'react-router-dom';
 import { TabBarContainer, TapBarItem } from './style';
 import { AiOutlineHome, AiOutlineGift, AiOutlineUser } from 'react-icons/ai';
 import { ROUTER_PATH } from '../../../../constants';
-import { RouterPath } from '../../../../types';
+import { RouterPath } from '../../../../types/utils';
 
 const BOTTOM_TABS = [
   { path: [ROUTER_PATH.couponList], icon: <AiOutlineHome size={28} />, label: '홈' },
