+++ frontend/src/pages/Admin/CouponDesign/ChoiceTemplate/index.tsx
@@ -5,9 +5,13 @@ import { TEMPLATE_MENU, TEMPLATE_OPTIONS } from '../../../../constants';
 import { useLocation } from 'react-router-dom';
 import { getCouponSamples } from '../../../../api/get';
 import { parseStampCount } from '../../../../utils';
-import { SampleBackCouponImage, SampleImage, StampCoordinate } from '../../../../types';
 import { useQuery } from '@tanstack/react-query';
-import { SampleCouponRes } from '../../../../types/api';
+import { SampleCouponRes } from '../../../../types/api/response';
+import {
+  StampCoordinate,
+  SampleImage,
+  SampleBackCouponImage,
+} from '../../../../types/domain/coupon';
 
 interface ChoiceTemplateProps {
   frontImageUrl: string;
