+++ frontend/src/pages/Admin/CouponDesign/CustomCouponDesign/StampCustomModal/index.tsx
@@ -3,7 +3,8 @@ import Modal from '../../../../../components/Modal';
 import { BackCouponWrapper, BackImage, ButtonContainer, Stamp, StampBadge } from './style';
 import Text from '../../../../../components/Text';
 import Button from '../../../../../components/Button';
-import { Coordinate, StampCoordinate } from '../../../../../types';
+import { StampCoordinate } from '../../../../../types/domain/coupon';
+import { Coordinate } from '../../../../../types/utils';
 
 interface Props {
   isOpen: boolean;
