+++ frontend/src/pages/Admin/ModifyCouponPolicy/CreatedType/index.tsx
@@ -9,9 +9,9 @@ import {
 } from './style';
 import { Spacing } from '../../../../style/layout/common';
 import Text from '../../../../components/Text';
-import { CouponCreated } from '../../../../types';
 import { GrSelect } from 'react-icons/gr';
 import { MdOutlinePhotoSizeSelectLarge } from 'react-icons/md';
+import { CouponCreated } from '../../../../types/domain/coupon';
 
 interface CreatedTypeProps {
   value: CouponCreated;
