+++ frontend/src/pages/Admin/RegisterCafe/index.tsx
@@ -7,8 +7,8 @@ import { useNavigate } from 'react-router-dom';
 import useFindAddress from '../../../hooks/useFindAddress';
 import { postRegisterCafe } from '../../../api/post';
 import { ROUTER_PATH } from '../../../constants';
-import { CafeRegisterReqBody, MutateReq } from '../../../types/api';
 import { useCafeQuery } from '../../../hooks/useRedirectRegisterPage';
+import { CafeRegisterReqBody } from '../../../types/api/request';
 
 const RegisterCafe = () => {
   const businessRegistrationNumberInputRef = useRef<HTMLInputElement>(null);
