+++ frontend/src/pages/Customer/HistoryPage/RewardHistory/index.tsx
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import { getMyRewards } from '../../../../api/get';
 import { CafeName, HistoryItem, DateTitle, HistoryList } from '../style';
 import { parseStringDateToKorean, sortMapByKey, transformEntries } from '../../../../utils';
-import { RewardHistoryDateProperties, RewardHistoryType } from '../../../../types';
+import { RewardHistoryDateProperties, RewardHistoryType } from '../../../../types/domain/reward';
 import { DATE_PARSE_OPTION } from '../../../../constants';
 import HistoryPage from '../HistoryPage';
 import CustomerLoadingSpinner from '../../../../components/LoadingSpinner/CustomerLoadingSpinner';
