+++ frontend/src/types/domain/cafe.ts
@@ -1,0 +1,11 @@
+export interface Cafe {
+  id: number;
+  name: string;
+  introduction: string;
+  openTime: string;
+  closeTime: string;
+  telephoneNumber: string;
+  cafeImageUrl: string;
+  roadAddress: string;
+  detailAddress: string;
+}
