+++ frontend/src/types/domain/customer.ts
@@ -1,0 +1,23 @@
+export interface Customer {
+  id: number;
+  nickname: string;
+  stampCount: number;
+  maxStampCount: number;
+  rewardCount: number;
+  visitCount: number;
+  firstVisitDate: string;
+  isRegistered: boolean;
+}
+
+export interface CustomerPhoneNumber {
+  id: number;
+  nickname: string;
+  phoneNumber: string;
+}
+
+export interface CustomerProfile {
+  id: number;
+  nickname: string;
+  phoneNumber: string | null;
+  email: string | null;
+}
