+++ frontend/src/types/domain/stamp.ts
@@ -1,0 +1,6 @@
+export interface StampHistoryType {
+  id: number;
+  cafeName: string;
+  stampCount: number;
+  createdAt: string;
+}
